made the changes in the footer
I started my web page by creating a new HTML file named index.html and writing the basic structure of an HTML document. This included the <!DOCTYPE html> declaration, the <html>, <head>, and <body> sections, as well as a <title> for the page and some simple content like a heading (<h1>) and a paragraph (<p>). After saving the file, I opened it in a web browser to view my first web page. This exercise helped me understand the fundamental structure of an HTML page and set the foundation for adding more content, styling with CSS, and interactivity with JavaScript in the future.